1、完成security模块的开发

+++ irh-service/irh-service-goods/src/main/java/top/imuster/goods/dao/ProductRecommendDao.java
@@ -14,5 +14,5 @@ import top.imuster.goods.api.dto.ProductRecommendDto;
 public interface ProductRecommendDao extends MongoRepository<ProductRecommendDto, String> {
     ProductRecommendDto findByUserId(Long userId);
 
-    ProductRecommendDto findProductRecommendDtoByUserId(Integer userId);
+    ProductRecommendDto findProductRecommendDtoByUserId(Long userId);
 }
